+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Enhancements
+
+- Improve `Ethers.deploy/2` error handling
+- Implement `Ethers.CcipRead` to support EIP-3668
+
 ## v0.5.5 (2024-12-03)
 
 ### Enhancements